requirements.txt
anyio==3.5.0
appdirs==1.4.4
asgiref==3.5.0
black==21.7b0
cfgv==3.3.1
click==8.0.4
colorama==0.4.4
distlib==0.3.4
dnspython==2.2.0
fastapi==0.74.1
filelock==3.6.0
h11==0.13.0
httptools==0.3.0
identify==2.4.10
idna==3.3
isort==5.5.2
mypy-extensions==0.4.3
nodeenv==1.6.0
pathspec==0.9.0
platformdirs==2.5.1
pre-commit==2.15.0
pydantic==1.9.0
pymongo==4.0.1
python-decouple==3.4
python-dotenv==0.19.2
PyYAML==6.0
regex==2022.1.18
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
toml==0.10.2
tomli==1.2.3
typing-extensions==4.1.1
uvicorn==0.17.5
virtualenv==20.13.1
watchgod==0.7
websockets==10.2
